+++ manager-api/src/core/fail2ban/fail2ban.module.ts
@@ -1,8 +1,9 @@
 import { Module } from "@nestjs/common";
+import { Fail2banDbService } from "./fail2ban-db.service";
 import { Fail2banService } from "./fail2ban.service";
 
 @Module({
-  providers: [Fail2banService],
+  providers: [Fail2banDbService, Fail2banService],
   exports: [Fail2banService],
 })
 export class Fail2banCoreModule {}